moving handlers and routes into backend
