+++ public/templates/dashboard.html
@@ -2,9 +2,10 @@
 <html>
 	<head>
 		<meta charset='utf-8'>
-		<title>Homepage</title>
+		<title>Dashboard page</title>
 	</head>
 	<body>
+    <h1>Dashboard page</h1>
 		<h1>Hello {{name}}</h1>
     <h3><a href='/logout'>Logout</a></h3>
     <h3><a href='/profile'>Profile</a></h3>
