+++ public/templates/index.html
@@ -5,7 +5,7 @@
 		<title>Landing page</title>
 	</head>
 	<body>
-		<h1>Hello {{name}}</h1>
+		<h1>Landing page</h1>
 		<h3><a href='/login'>Login</a></h3>
 	</body>
 </html>
